Reduce samples
Instead of generating samples for all notes, which requires too much memory in the ESP32, we only generate samples for the first octave. Notes for all other octaves are generated by increasing the step through the samples - for octave 1, we use a step of 2, octave 2, step of 4, and so on.
Could be even further reduced by, say, generating notes for octave 2 and creating samples for octaves 0 and 1 by interpolation. Ugly sound, probably.